+++ src/biodwh2-pharmgkb/src/main/java/de/unibi/agbi/biodwh2/pharmgkb/etl/PharmGKBGraphExporter.java
@@ -48,7 +48,7 @@ public class PharmGKBGraphExporter extends GraphExporter<PharmGKBDataSource> {
 
     @Override
     public long getExportVersion() {
-        return 4;
+        return 5;
     }
 
     @Override