+++ src/biodwh2-pharmgkb/src/main/java/de/unibi/agbi/biodwh2/pharmgkb/model/Chemical.java
@@ -3,8 +3,8 @@ package de.unibi.agbi.biodwh2.pharmgkb.model;
 import com.univocity.parsers.annotations.Parsed;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphArrayProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphBooleanProperty;
-import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphNodeLabel;
+import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 
 @GraphNodeLabel("Chemical")
 public class Chemical {
@@ -15,19 +15,19 @@ public class Chemical {
     @GraphProperty("name")
     public String name;
     @Parsed(field = "Generic Names")
-    @GraphArrayProperty(value = "generic_names", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "generic_names", arrayDelimiter = ", ", quotedArrayElements = true)
     public String genericNames;
     @Parsed(field = "Trade Names")
-    @GraphArrayProperty(value = "trade_names", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "trade_names", arrayDelimiter = ", ", quotedArrayElements = true)
     public String tradeNames;
     @Parsed(field = "Brand Mixtures")
-    @GraphArrayProperty(value = "brand_mixtures", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "brand_mixtures", arrayDelimiter = ", ", quotedArrayElements = true)
     public String brandMixtures;
     @Parsed(field = "Type")
-    @GraphArrayProperty(value = "types", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "types", arrayDelimiter = ", ", quotedArrayElements = true)
     public String type;
     @Parsed(field = "Cross-references")
-    @GraphArrayProperty(value = "cross_references", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "cross_references", arrayDelimiter = ", ", quotedArrayElements = true)
     public String crossReferences;
     @Parsed(field = "SMILES")
     @GraphProperty("smiles")
@@ -39,7 +39,7 @@ public class Chemical {
     @GraphBooleanProperty(value = "dosing_guideline", truthValue = "yes")
     public String dosingGuideline;
     @Parsed(field = "External Vocabulary")
-    @GraphArrayProperty(value = "external_vocabulary", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "external_vocabulary", arrayDelimiter = ", ", quotedArrayElements = true)
     public String externalVocabulary;
     @Parsed(field = "Clinical Annotation Count")
     @GraphProperty("clinical_annotation_count")
@@ -54,7 +54,7 @@ public class Chemical {
     @GraphProperty("vip_count")
     public Integer vipCount;
     @Parsed(field = "Dosing Guideline Sources")
-    @GraphArrayProperty(value = "dosing_guideline_sources", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "dosing_guideline_sources", arrayDelimiter = ", ", quotedArrayElements = true)
     public String dosingGuidelineSources;
     @Parsed(field = "Top Clinical Annotation Level")
     @GraphProperty("top_clinical_annotation_level")
@@ -72,12 +72,12 @@ public class Chemical {
     @GraphProperty("has_rx_annotation")
     public String hasRxAnnotation;
     @Parsed(field = "RxNorm Identifiers")
-    @GraphArrayProperty(value = "rxnorm_identifiers", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "rxnorm_identifiers", arrayDelimiter = ", ", quotedArrayElements = true)
     public String rxNormIdentifiers;
     @Parsed(field = "ATC Identifiers")
-    @GraphArrayProperty(value = "atc_identifiers", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "atc_identifiers", arrayDelimiter = ", ", quotedArrayElements = true)
     public String atcIdentifiers;
     @Parsed(field = "PubChem Compound Identifiers")
-    @GraphArrayProperty(value = "pubchem_compound_identifiers", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "pubchem_compound_identifiers", arrayDelimiter = ", ", quotedArrayElements = true)
     public String pubChemCompoundIdentifiers;
 }