+++ src/biodwh2-pharmgkb/src/main/java/de/unibi/agbi/biodwh2/pharmgkb/model/Gene.java
@@ -3,8 +3,8 @@ package de.unibi.agbi.biodwh2.pharmgkb.model;
 import com.univocity.parsers.annotations.Parsed;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphArrayProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphBooleanProperty;
-import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphNodeLabel;
+import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 
 @GraphNodeLabel("Gene")
 public class Gene {
@@ -12,13 +12,13 @@ public class Gene {
     @GraphProperty("id")
     public String pharmgkbAccessionId;
     @Parsed(field = "NCBI Gene ID")
-    @GraphArrayProperty(value = "ncbi_gene_ids", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "ncbi_gene_ids", arrayDelimiter = ", ")
     public String ncbiGeneId;
     @Parsed(field = "HGNC ID")
-    @GraphArrayProperty(value = "hgnc_ids", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "hgnc_ids", arrayDelimiter = ", ")
     public String hgncId;
     @Parsed(field = "Ensembl Id")
-    @GraphArrayProperty(value = "ensemble_ids", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "ensemble_ids", arrayDelimiter = ", ")
     public String ensembleId;
     @Parsed(field = "Name")
     @GraphProperty("name")
@@ -27,10 +27,10 @@ public class Gene {
     @GraphProperty("symbol")
     public String symbol;
     @Parsed(field = "Alternate Names")
-    @GraphArrayProperty(value = "alternate_names", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "alternate_names", arrayDelimiter = ", ", quotedArrayElements = true)
     public String alternateNames;
     @Parsed(field = "Alternate Symbols")
-    @GraphArrayProperty(value = "alternate_symbols", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "alternate_symbols", arrayDelimiter = ", ", quotedArrayElements = true)
     public String alternateSymbols;
     @Parsed(field = "Is VIP")
     @GraphBooleanProperty(value = "is_vip", truthValue = "yes")
@@ -39,7 +39,7 @@ public class Gene {
     @GraphBooleanProperty(value = "has_variant_annotation", truthValue = "yes")
     public String hasVariantAnnotation;
     @Parsed(field = "Cross-references")
-    @GraphArrayProperty(value = "cross_references", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "cross_references", arrayDelimiter = ", ", quotedArrayElements = true)
     public String crossReference;
     @Parsed(field = "Has CPIC Dosing Guideline")
     @GraphBooleanProperty(value = "has_cpic_dosing_guideline", truthValue = "yes")