+++ src/biodwh2-pharmgkb/src/main/java/de/unibi/agbi/biodwh2/pharmgkb/model/Phenotype.java
@@ -2,8 +2,8 @@ package de.unibi.agbi.biodwh2.pharmgkb.model;
 
 import com.univocity.parsers.annotations.Parsed;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphArrayProperty;
-import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphNodeLabel;
+import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 
 @GraphNodeLabel("Phenotype")
 public class Phenotype {
@@ -14,12 +14,12 @@ public class Phenotype {
     @GraphProperty("name")
     public String name;
     @Parsed(field = "Alternate Names")
-    @GraphArrayProperty(value = "alternate_names", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "alternate_names", arrayDelimiter = ", ", quotedArrayElements = true)
     public String alternateNames;
     @Parsed(field = "Cross-references")
-    @GraphArrayProperty(value = "cross_references", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "cross_references", arrayDelimiter = ", ", quotedArrayElements = true)
     public String crossReferences;
     @Parsed(field = "External Vocabulary")
-    @GraphArrayProperty(value = "external_vocabulary", arrayDelimiter = ",", quotedArrayElements = true)
+    @GraphArrayProperty(value = "external_vocabulary", arrayDelimiter = ", ", quotedArrayElements = true)
     public String externalVocabulary;
 }